% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessSEM.assignDS.R
\name{hessSEM.assignDS}
\alias{hessSEM.assignDS}
\title{Concatenates objects into a vector or list}
\usage{
hessSEM.assignDS(x)
}
\arguments{
\item{x}{a list which contains the the objects to concatenate.}
}
\value{
a vector or list
}
\description{
This function is similar to the R base function 'c'.
}
\details{
Unlike the R base function 'c' on vector or list of certain
}
\author{
DataSHIELD Development Team
}
