% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessSEMDS.R
\name{hessSEMDS}
\alias{hessSEMDS}
\title{Concatenates objects into a vector or list}
\usage{
hessSEMDS(x, chains, method, seed, iter)
}
\arguments{
\item{x}{a list which contains the the objects to concatenate.}

\item{chains}{is the ...}

\item{method}{is the ...}

\item{seed}{is the ...}

\item{iter}{is the ...}
}
\value{
a vector or list
}
\description{
This function is similar to the R base function 'c'.
}
\details{
Unlike the R base function 'c' on vector or list of certain
}
\author{
DataSHIELD Development Team
}
