% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessSEMDS.R
\name{hessSEMDS}
\alias{hessSEMDS}
\title{Concatenates objects into a vector or list}
\usage{
hessSEMDS(
  resource,
  infile,
  autoAddIntercept,
  gammaInit,
  nIter,
  burnin,
  nChains,
  seed,
  method
)
}
\arguments{
\item{resource}{a resource which is used to access the data set.}

\item{infile}{is the ...}

\item{autoAddIntercept}{is the ...}

\item{gammaInit}{is the ...}

\item{nIter}{is the ...}

\item{burnin}{is the ...}

\item{nChains}{is the ...}

\item{seed}{is the ...}

\item{method}{is the ...}
}
\value{
a vector or list
}
\description{
This function is similar to the R base function 'c'.
}
\details{
Unlike the R base function 'c' on vector or list of certain
}
\author{
DataSHIELD Development Team
}
